Add unit tests for Mamdani defuzzification and implement routine logic validation
- Created unit tests for Mamdani defuzzification in `test_mamdani_defuzzification.py`, covering various scenarios including actuator logic for PWM and DIGITAL variables, membership function creation, and validation of FuzzyVariable properties.
- Implemented `validate_routine_logic.py` to validate routine logic in the fuzzy service, checking for inconsistencies between actuator types and command types, redundancy in steps, and ensuring unique output variables.
- Added `verify_seed_conversion.py` to test the conversion of routines to consequents, ensuring correct mapping of terms and validation of generated consequents.

+++ software-project/fuzzy-service/FuzzyService/Application/Features/ActuatorIntegration/Handlers/SendRoutinesToActuatorHandler.py
@@ -50,14 +50,14 @@ class SendRoutinesToActuatorHandler(CommandHandler[SendRoutinesToActuatorCommand
                 for step in routine.steps:
                     step_dict = {
                         "outputVariable": step.outputVariable,
-                        "duration": int(step.duration)
+                        "duration": float(step.duration)  # Mantener como float, no int
                     }
 
                     # Agregar power O dutyCycle (mutuamente exclusivos)
                     if step.power is not None:
                         step_dict["power"] = step.power
                     if step.dutyCycle is not None:
-                        step_dict["dutyCycle"] = step.dutyCycle
+                        step_dict["dutyCycle"] = float(step.dutyCycle)  # También float
 
                     routine_dict["steps"].append(step_dict)
